third_party/blink/web_tests/external/wpt/html/cross-origin-embedder-policy/reflection-require-corp.tentative.https.any.worker-expected.txt

This is a testharness.js-based test.
[FAIL] reflection-require-corp
  assert_equals: expected "require-corp" but got "unsafe-none"
Harness: the test ran to completion.
